+++ lib/apollovm.dart
@@ -5,4 +5,4 @@ export 'src/apollovm_parser.dart';
 export 'src/apollovm_ast.dart';
 
 export 'src/languages/dart/dart_parser.dart';
-export 'src/languages/java/java8/java8_parser.dart';
+export 'src/languages/java/java11/java11_parser.dart';