+++ lib/src/languages/java/java11/java11_grammar.dart
@@ -2,10 +2,10 @@ import 'package:apollovm/apollovm.dart';
 import 'package:apollovm/src/apollovm_ast.dart';
 import 'package:petitparser/petitparser.dart';
 
-import 'java8_grammar_lexer.dart';
+import 'java11_grammar_lexer.dart';
 
-/// Java8 grammar definition.
-class Java8GrammarDefinition extends Java8GrammarLexer {
+/// Java11 grammar definition.
+class Java11GrammarDefinition extends Java11GrammarLexer {
   static ASTType getTypeByName(String name) {
     switch (name) {
       case 'Object':
@@ -90,7 +90,7 @@ class Java8GrammarDefinition extends Java8GrammarLexer {
       });
 
   Parser<ASTStatement> statement() =>
-      (statementVariableDeclaration() | statementExpression())
+      (branch() | statementVariableDeclaration() | statementExpression())
           .cast<ASTStatement>();
 
   Parser<ASTStatementExpression> statementExpression() =>
@@ -107,6 +107,67 @@ class Java8GrammarDefinition extends Java8GrammarLexer {
         var valueOpt = v[2];
         var value = valueOpt != null ? valueOpt[1] : null;
         return ASTStatementVariableDeclaration(v[0], v[1], value);
+      });
+
+  Parser<ASTBranch> branch() => (ref0(branchIfElseIfsElseBlock) |
+          ref0(branchIfElseBlock) |
+          ref0(branchIfBlock))
+      .cast<ASTBranch>();
+
+  Parser<ASTBranchIfBlock> branchIfBlock() => (string('if').trim() &
+              char('(').trim() &
+              ref0(expression) &
+              char(')').trim() &
+              codeBlock())
+          .map((v) {
+        var condition = v[2];
+        var block = v[4];
+        return ASTBranchIfBlock(condition, block);
+      });
+
+  Parser<ASTBranchIfElseBlock> branchIfElseBlock() => (string('if').trim() &
+              char('(').trim() &
+              ref0(expression) &
+              char(')').trim() &
+              codeBlock() &
+              string('else').trim() &
+              codeBlock())
+          .map((v) {
+        var condition = v[2];
+        var blockIf = v[4];
+        var blockElse = v[6];
+        return ASTBranchIfElseBlock(condition, blockIf, blockElse);
+      });
+
+  Parser<ASTBranchIfElseIfsElseBlock> branchIfElseIfsElseBlock() =>
+      (string('if').trim() &
+              char('(').trim() &
+              ref0(expression) &
+              char(')').trim() &
+              codeBlock() &
+              ref0(branchElseIfs).plus() &
+              string('else').trim() &
+              codeBlock())
+          .map((v) {
+        var condition = v[2];
+        var blockIf = v[4];
+        var blockElseIfs = v[5] as List;
+        var blockElse = v[7];
+
+        return ASTBranchIfElseIfsElseBlock(condition, blockIf,
+            blockElseIfs.cast<ASTBranchIfBlock>().toList(), blockElse);
+      });
+
+  Parser<ASTBranchIfBlock> branchElseIfs() => (string('else').trim() &
+              string('if').trim() &
+              char('(').trim() &
+              ref0(expression) &
+              char(')').trim() &
+              codeBlock())
+          .map((v) {
+        var condition = v[3];
+        var blockIf = v[5];
+        return ASTBranchIfBlock(condition, blockIf);
       });
 
   Parser<ASTExpression> expression() => (ref0(expressionNoOperation) &
@@ -140,7 +201,11 @@ class Java8GrammarDefinition extends Java8GrammarLexer {
               char('*') |
               char('/') |
               string('==') |
-              string('!='))
+              string('!=') |
+              string('<=') |
+              string('>=') |
+              char('<') |
+              char('>'))
           .trim()
           .map((v) {
         return getASTExpressionOperator(v);